Basics for Size of Anything
You can speak or type the name of any place, and it appears on the map. Then, you can drag it around the map. It does not appear in the Active Element Panel (there's no semblence of "active"), you can only have one element at a time, and the discoveries panel is completely unimplemented. There is very limited styling.